+++ zebrad/src/commands/start.rs
@@ -39,7 +39,6 @@ pub struct StartCmd {
 impl StartCmd {
     async fn start(&self) -> Result<(), Report> {
         info!(?self, "starting to connect to the network");
-        tokio::spawn(on_ctrl_c());
 
         let config = app_config();
         let state = zebra_state::on_disk::init(config.state.clone(), config.network.network);
@@ -59,12 +58,6 @@ impl StartCmd {
 
         syncer.sync().await
     }
-}
-
-async fn on_ctrl_c() {
-    tokio::signal::ctrl_c().await.unwrap();
-    let _ = crate::application::FLAME_GUARD.lock().unwrap().take();
-    std::process::exit(1);
 }
 
 impl Runnable for StartCmd {
